qtx_fishTest: split up integration and unit tests
