+++ engine/NQ/tests/fishtests.h
@@ -1,0 +1,12 @@
+#ifndef FISHTESTS_H
+#define FISHTESTS_H
+
+void test_ray_to_latlon(void **state);
+
+void test_latlon_to_ray(void **state);
+
+void test_plate_uv_to_ray(void **state);
+
+void test_uv_to_screen(void **state);
+
+#endif
